Remove warning in JPA
