+++ app/views/index.scala.html
@@ -5,7 +5,7 @@
 
     <ul id="persons"></ul>
 
-    <form method="POST" action="@routes.Application.addPerson()">
+    <form method="POST" action="@routes.PersonController.addPerson()">
         <input type="text" name="name"/>
         <button>Add Person</button>
     </form>
