+++ build.sbt
@@ -4,12 +4,14 @@ version := "1.0-SNAPSHOT"
 
 lazy val root = (project in file(".")).enablePlugins(PlayJava)
 
-scalaVersion := "2.11.7"
+scalaVersion := "2.11.8"
 
-libraryDependencies ++= Seq(
-  // If you enable PlayEbean plugin you must remove these
-  // JPA dependencies to avoid conflicts.
-  javaJpa,
-  "org.hibernate" % "hibernate-entitymanager" % "4.3.7.Final"
-)
+libraryDependencies += javaJpa
 
+// https://mvnrepository.com/artifact/dom4j/dom4j
+libraryDependencies += "dom4j" % "dom4j" % "1.6"
+
+// https://mvnrepository.com/artifact/org.hibernate/hibernate-core
+// must exclude dom4j in hibernate core because it causes staxeventreader exceptions
+// http://stackoverflow.com/questions/36222306/caused-by-java-lang-classnotfoundexception-org-dom4j-io-staxeventreader
+libraryDependencies += "org.hibernate" % "hibernate-core" % "5.2.3.Final" exclude("dom4j", "dom4j") exclude("javax.transaction", "jta") exclude("org.slf4j", "slf4j-api")
